+++ lib/sqlalchemy/util/concurrency.py
@@ -1,3 +1,10 @@
+# util/concurrency.py
+# Copyright (C) 2005-2021 the SQLAlchemy authors and contributors
+# <see AUTHORS file>
+#
+# This module is part of SQLAlchemy and is released under
+# the MIT License: https://www.opensource.org/licenses/mit-license.php
+
 from . import compat
 
 have_greenlet = False
@@ -19,7 +26,9 @@ if compat.py3k:
             _util_async_run_coroutine_function,
         )  # noqa F401, E501
         from ._concurrency_py3k import asyncio  # noqa F401
-        from ._concurrency_py3k import asynccontextmanager
+
+    # does not need greennlet, just Python 3
+    from ._compat_py3k import asynccontextmanager  # noqa F401
 
 if not have_greenlet:
 
@@ -57,7 +66,4 @@ if not have_greenlet:
         return fn(*arg, **kw)
 
     def _util_async_run_coroutine_function(fn, *arg, **kw):  # noqa F81
-        _not_implemented()
-
-    def asynccontextmanager(fn, *arg, **kw):  # noqa F81
         _not_implemented()
